archive-2026-03-10T165238Z/requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
jinja2>=3.1.0
flask>=3.0.0
motor>=3.3.0
pymongo>=4.6.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
aiohttp>=3.9.0
Pillow>=10.0.0
discord.py>=2.3.0
yt-dlp>=2024.1.0
PyNaCl>=1.5.0
requests>=2.31.0
aiofiles>=23.0.0

